backend/utils/image_processor.py
import base64
import os
import uuid
import requests
from PIL import Image
import io

def process_image(base64_image):
    """
    Process an image and return a description.
    
    This function uses a simple approach to describe basic image attributes.
    In a production environment, you would integrate with an AI vision model
    like Azure Computer Vision, Google Cloud Vision, or similar.
    
    Args:
        base64_image (str): Base64-encoded image data
    
    Returns:
        str: Description of the image
    """
    try:
        # Extract the base64 content (remove data:image/jpeg;base64, prefix)
        if ',' in base64_image:
            base64_image = base64_image.split(',')[1]
        
        # Decode base64 to binary
        image_data = base64.b64decode(base64_image)
        
        # Create a temporary file
        temp_image_path = f"temp_image_{uuid.uuid4()}.jpg"
        
        # Save the image temporarily
        with open(temp_image_path, 'wb') as img_file:
            img_file.write(image_data)
        
        # Open the image with PIL
        img = Image.open(temp_image_path)
        
        # Get basic image information
        width, height = img.size
        format_name = img.format
        mode = img.mode
        
        # Since we don't have direct access to image analysis APIs here,
        # we'll return basic information and a placeholder for real analysis
        
        # If in production, this is where you would call a vision API
        # For example:
        # description = call_vision_api(temp_image_path)
        
        description = (
            f"This is an image in {format_name} format with dimensions {width}x{height} pixels. "
            f"The color mode is {mode}. "
            "(Note: In a production environment, this would include a detailed description "
            "of the image content from a vision API.)"
        )
        
        return description
        
    except Exception as e:
        return f"Error processing image: {str(e)}"
        
    finally:
        # Clean up
        if 'temp_image_path' in locals() and os.path.exists(temp_image_path):
            os.remove(temp_image_path)