backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
groq==0.4.0
python-dotenv==1.0.0
pillow==10.4.0
SpeechRecognition==3.10.0
pydub==0.25.1
requests==2.31.0
uuid==1.30